+++ cmd/certkit/spinner.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"os"
 	"sync"
-	"sync/atomic"
 	"time"
 
 	"github.com/mattn/go-isatty"
@@ -20,7 +19,6 @@ type spinner struct {
 	done      chan struct{}
 	startOnce sync.Once
 	stopOnce  sync.Once
-	started   atomic.Bool
 }
 
 // newSpinner creates a spinner with the given message. Call Start() to begin
@@ -44,7 +42,6 @@ func (s *spinner) Start(ctx context.Context) {
 			return
 		}
 
-		s.started.Store(true)
 		go s.run(ctx)
 	})
 }
@@ -58,10 +55,6 @@ func (s *spinner) SetMessage(msg string) {
 
 // Stop halts the spinner and clears the line. Safe to call multiple times.
 func (s *spinner) Stop() {
-	if !s.started.Load() {
-		<-s.done
-		return
-	}
 	s.stopOnce.Do(func() { close(s.stop) })
 	<-s.done
 }